顧客_レビュー削除機能

+++ db/migrate/20240430122317_create_reviews.rb
@@ -4,11 +4,11 @@ class CreateReviews < ActiveRecord::Migration[6.1]
       t.integer :customer_id, null: false
       t.integer :facility_id, null: false
       t.string :title, null: false
-      t.integer :total_score, null: false
-      t.integer :facility_score, null: false
-      t.integer :onsen_score, null: false
-      t.integer :service_score, null: false
-      t.integer :meal_score, null: false
+      t.float :total_score, null: false
+      t.float :facility_score, null: false
+      t.float :onsen_score, null: false
+      t.float :service_score, null: false
+      t.float :meal_score, null: false
       t.text :comment, null: false
       t.boolean :is_active, default: true, null: false
