+++ db/schema.rb
@@ -129,11 +129,11 @@ ActiveRecord::Schema.define(version: 2024_05_31_115134) do
     t.integer "customer_id", null: false
     t.integer "facility_id", null: false
     t.string "title", null: false
-    t.integer "total_score", null: false
-    t.integer "facility_score", null: false
-    t.integer "onsen_score", null: false
-    t.integer "service_score", null: false
-    t.integer "meal_score", null: false
+    t.float "total_score", null: false
+    t.float "facility_score", null: false
+    t.float "onsen_score", null: false
+    t.float "service_score", null: false
+    t.float "meal_score", null: false
     t.text "comment", null: false
     t.boolean "is_active", default: true, null: false
     t.datetime "created_at", precision: 6, null: false
